Add downloaded files to the media library.
Since we can't download the file directly to it's final destination, we can't
have the download manager requisting the file scanned by the media server.
Instead we do it in broadcast receiver which listens for completed downloads.
As it is not allowed to bind to a service from a broadcast receiver, we can't
call MediaScannerConnection.scanFile, because it internally binds to the
media service.
So we use the ACTION_MEDIA_SCANNER_SCAN_FILE intent action to have the file
scanned.

This fixes a comment on issue #55.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>