Update the UI when player states change, not players.
As the service is tracking players from the time it connects to the
server there is no need to re-fetch the list of players in this
activity.

Instead, use the service's current list of players, and rely on player
state change notifications to update the UI.

Refactor some method names and variables to make more sense.