Create and use an EventBus that's specific to the service.
Instead of using the default EventBus provided by EventBus.getDefault(),
SqueezeService creates its own, and makes it available to clients of
the service through the getEventBus() method.

This ensures that service events are only sent to clients of the service
instead of being received by all-comers, which will probably be useful
later.

This changes the lifecycle of the activity / fragment registration on
the event bus. Instead of using onStart() and onStop() to register,
register in maybeRegisterCallbacks(), and unregister in onPause(),
to mirror when the callbacks are registered.