Fix timepicker on honeycomb
It appears that onDraw is not called on honeycomb, for the radial views,
which are not visible. This causes the animators to not be ready, when
they are needed when the radial picker layout requests a change of the
displayed radial views (hours <-> minutes), as they animators are rendered
in onDraw.

To fix this, the animators are rendered on demand, and the requirement,
that onDraw has to be called before the animators are requested is removed.

We still prerender the animations in onDraw, so on all other platforms,
the timepicker should function unchanged.