Fix HandshakeNotCompleteExceptions by moving setAdapter() calls.
Calling setAdapter triggers a layout measure pass, and the measure
pass ends up calling getView(), which can call through to the service
to request the album art URL, which only works after the handshake
is complete.

Where setAdapater() calls were happening in the onCreate() or
onServiceConnected() methods, create a handshake callback and set the
adapter there.