Wrap the trackname TextView in a LinearLayout.
Without this, changing the contents of any other item in the layout will
cause the layout to reflow, which causes the scrolling marquee text to
reset.  Wrapping the trackname TextView in a LinearLayout means that the
TextView will not change size when then the layout is recalculated, and
the text continues to scroll.