Ignore context menu selections if players change while the menu is shown.
This replaces the previous check which made sure the list of players
was not empty.

This more general check covers situations where players come and go
from the list and the context menu being displayed is for a player that
is no longer connected, or that has moved position in the list.

Fix for http://crashes.to/s/243a873f36b.