Warm welcome - broadcast for local servers, and
let the user choose one, and optionally set credentials.

When starting Squeezer for the first time, don't dump people in the settings
straight off.

To be able to start the server address dialog, without first going to settings,
the dialog can no longer extend DialogPreference. To still be able to set the
prefered server via settings, and to avoid code and resource duplication, I have
created a custom view with all the functionality, which can be used from both.
Also the local servers are searched without waiting for the user to request it.