Show disconnected activity on login failure
with an additional informational header explaining the issue, and a link to an
alart dialog which explains how to check the server settings.

If the server disconnects before the handshake is completed, it was interpreted
as an authentification failure. If the server connects but then disconnects for
some other reason before authentication is done, Squeezer would wrongly ask for
authentication. This happens e.g. if you enter the web interface port instead
of the CLI port.

The disconnected activity (via the server address view) gives the possibility
to change all the connection parameters, and the informational tekst does not
assume that it is an authentication failure.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>