Fix a bug determining the correct player subscription type.
The player subscription type (real_time, etc) is determined by the
events that clients of mEventBus are registered for.

However, clients that register after the last call to
updateAllPlayerSubscriptionStates() are not considered.

This manifests as starting Squeezer connected to a server with a player
that is already playing a track.  Because NowPlayingFragment registers
with the service after updateAllPlayerSubscriptionStates() is called,
which means that the logic that checks for clients subscribed to the
SongTimeChanged event fails, and the player's subscription type is
not set to 'real_time'.

The result is that the progress bar in NowPlayingFragment does not
progress through the track, in both the mini player and the full screen
player modes.

Fix this by introducing a subclass of EventBus that calls through to
updateAllPlayerSubscriptionStates() every time a client registers or
unregisters on the bus.

While in this code, renamed getPlayerSubscriptionType() to
calculateSubscriptionTypeFor() to make it clearer that this method
does not simply return a static value.