Workaround for internet radios not working on 7.8+
Internet radio stations can't be parsed on recent servers. According to the CLI
documentation the "cmd" column is suppose to be the item delimiter for the
"radios" command. However the columns arive in random order, so when the cmd
column arrives, the previous record may not be complete.

To workaround this, we test that all columns for a radio has arrived before we
proceed with the next entry.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>