Fix icon handling
Some plugins return an absolute url for the icons, so we test if it's absolute
before we add the local server to the path.

Also some servers doesn't like an extra slash, e.g.: "http://server:port//path/icon.png"
so we only add a slash, if the releative url doesn't already start with a slash.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>