Make play list affecting events sticky.
This affects MusicChanged, PlaylistTracksAdded, PlaylistTracksDeleted.

This means that opening up a playlist, leaving Squeezer while the
music continues to play, and coming back to the same activity through
the "Recents" menu restores the correct place in the playlist, instead
of sticking to whatever the last track that was played is.