Ensure ImageFetcher threads resume when the activity is paused.
Without this, an ImageFetcher thread could be spawned, paused (because
the user scrolls) and then the activity goes away because the user
presses the "back" button (or similar).

The thread is now running waiting for a notification that will never
come.

If this happens twice then the thread pool is filled with non-responsive
threads, and no more images can be loaded.

In addition, call setExitTasksEarly in onPause (enable it) and onResume
(disable it) so that the bitmap threads do not carry out unnecessary
work when the activity is pausing.