Fix a crash caused by showing dialogs when the fragment is not added.
The previous code called ProgressDialog.show() in startVisibleConnection()
which would occasionally crash.

Fix this by making startVisibleConnection() only responsible for
starting the connection, remove the code that creates the connecting
dialog.

Responsibility for showing and hiding dialogs, as well as the rest of
the UI changes, moves in to the onEventMainThread(ConnectionChanged) and
onEventMainThread(HandshakeComplete) methods. This removes the need
for setConnected() and updateUIFromServiceState().

Update ConnectionState so that LOGIN_FAILED is a terminal state. Without
this, the state transition goes from LOGIN_FAILED to DISCONNECTED and
the UI tips in DisconnectedActivity do not show for this state. Adjust
the unit test accordingly.

Update DisconnectedActivity to take an @IntDef that specifies the reason
for the disconnection (instead of a boolean). Provide appropriate helpful
text for these events. Only launch HomeActivity after the handshake
has completed.