Remove all "active" player handling from CliClient.
Move the code that deals with selecting and maintaining the active
player in to SqueezeService.

Update the events (which are predominantly posted by CliClient) to
remove references to the active player. Update the event handlers
that only operate on the active player to check that they are
receiving events about the active player.