Generate URLs for artwork, downloads, and icons when items are loaded.
The previous code used a "just in time" method to determine URLs for
an item's URLs for artwork and downloads. This required a round-trip
to SqueezeService and a dependency between model objects and
SqueezeService.

Break the dependency by ensuring that the tokens for "artwork_url",
"download_url", and "icon" are converted in to absolute URls as the
information is parsed from the server.