Correct how spinners are displayed.
Spinners have two views, one for the spinner "at rest" when it's
showing the selected item, and one that's used to construct the
list of items that the user can choose from. The previous code was
using the second version in both places, which caused issues with
the UI on earlier devices.

To fix this:

- Introduce SpinnerItemAdapter which overrides getDropDownView() to
  call through to the itemview's getDropDownAdapterView().

- Introduce SpinnerItemView which calls methods in the Util class
  to provide default implementations for itemviews that need it.

- Use SpinnerItemAdapter instead of ItemAdapter in GenreSpinner and
  YearSpinner.

- Change GenreView and YearView to inherit from SpinnerItemView.

- Stop trying to use something akin to a floatinglabel layout for
  the labels for spinners in AlbumFilterDialog, as it doesn't work
  properly across multiple API versions. Simplify with traditional
  labels, which also removes the need for fragile padding values on