Avoid network operations on the main thread when updating notifications.
- Refactor ImageWorker to provide additional loadImage(...) methods that
  can:
  - load an image, and then call a callback
  - load an image in to a RemoteViews, and post a notification

- Use these new methods in updateOngoingNotification() to load the
  images in to the notification's view and post the new notification.

This causes a problem with EventBus on pre-Lollipop devices, as it uses
Class#getDeclaredMethods() finds methods that do not exist or have
non-existent return types on earlier devices. Guarding the code with
@TargetApi or checks against Build.VERSION.SDK_INT does not solve
this problem.

This is discussed in issues on EventBus, such as
https://github.com/greenrobot/EventBus/issues/149#issuecomment-76316437.

The fork of EventBus at https://github.com/yuzeh/EventBus fixes this,
per the pull request at https://github.com/greenrobot/EventBus/pull/158
and https://github.com/yuzeh/EventBus/commit/dc326b0bc6966945924d616327ef0b657954562e
so for the moment pull that code in as a submodule and use it in
preference to the official EventBus code.