Reduce bitmap memory use.
The previous code would always decode the bitmap to full size in order
to save it to the disk cache. This could cause OOMs for massive bitmaps,
even if they were never going to be displayed at that size.

Change the code so that the disk cache now saves the image bytes that
were received (i.e., the undecoded bitmap). This keeps the full size
image on disk (encoded), so it needs to be redecoded every time its
fetched from disk, but that's relatively rare.