Implement popup for configurable actions
If a configurable action for song or album is not set. or is NONE, a
popup with the possible actions, and "Just once" and "Always" options,
similar to the standard Android way of resolving intents.

This requires Dialog.setOnShowListener so it is for API level 8 (froyo)
or later. API level 7 still have to select the default action via
settings.