Implement download options
This implements part of issue #132

The option to select Download storage (SDcard or internal) doesn't seem
to be possibe on recent Android version (4.4 or later), see for example:
"What about removable SD card in 4.4"
https://groups.google.com/forum/#!topic/android-platform/14VUiIgwUjY[1-25]

How the downloaded tracks are stored on the Android device is now user
selectable via preferences.

Default is to use the server path and filename, and this has been the only
option until now.

Otherwise path structure and file name structure can be constructed from
the properties of the downloaded tracks. The path and file name options
are modelled from the open source music player Rhythmbox, and the English
and Danish texts for the options are taken from there.

When tracks are downloaded via the Music Folder menu entry, it will be
stored using the server path and name, regardless of what is chosen in
the preferences.