Fix playlist position not initially updating if current title not loaded
Issue: when the currently playing track is far enough from the first view tracks that fit onto the playlist screen, and the playlist does not load all the other tracks, the current title is not selected until the user scrolls there manually.
After some experiments, I figured out that selectCurrentSong is called correctly, but the listView does not post the runnable for whatever reason.
Using a handler to post the update instead of the view fixes this