Workaround a bug in BayeuxClient
The BayeuxClient send a message id with the Subscribe Request and expect
it to be echoed in the Subscribe Response.

LMS doesn't do that, and as it is an optional field, it is according to spec.

BayeuxClient pass the message id to a ConcurrentHashMap which (at least on
Android) causes a NullPointerException.

To fix this we intercept the request to note the message id, and intercept
the repsonse to add the noted message id.