Handle responses with empty list when browsing
LMS items.

If there are no items count will be 0 and items_loop will not be
present in the response message, i.e. for the players request, it
means that players_loop will not be in the response, so we check
for null before trying to use the item data.