Implement commands
When CometClient.sendCommand sees a command to which there is not
defined a listener or a respinse channel, it as taken as a command
(instead of a request or a subscription), so it is just send to LMS
and no pending request is registered.

Apparently this means some unknown status messages comes in, so we
check whether it is a a message we have subscribed for, before parsing
it.