Implement current playlist request in CometClient
The status request which is used for both player status requests and
subscriptions, and to request the current playlist, requires some special
handling.

As the comet replies does not echo the request we set up different listeners
when we send the request.

To be able to determine which request it is we need the full command, so this
is passed to the methods which sets up the request and the message listeners.