Fix build problems
The comet client causes various problems with proguard. I have removed
warnings of unresolved classes in a number of packages which proguard
claims to ne needed.  Fingers crossed, that they are not really needed.
Also comment out the cometd-websocket-jetty dependency, it is not
currently used, and it brings in a whole new range of dependencis.

Also update repositories and build tools version in datetimepicker, and remove
JsonTest, as it depends on no longer used libraries.