Players now sorted by playername in drop down list and player setting page.
Some people on the slimdevices forum wondered if the player list can be
ordered alphabetically in Squeezer. I had noticed too that the player
list looked random (I now know it was ordered on mac-address).
I have some experience with java / Android, so I gave it a try. I got it
working.
The list in de player select dropdown list is now ordered and the list
in the settings player page is now ordered too (first on sync group, and
then the children within the group).
In my opinion, Squeezer is the best LMS controller app. I would be
honored if this change was included in a new Squeezer release.