Fix timeout issue in CometClient
When there are no events from LMS (and no active player state subscription)
we would get a TimeoutException and consequently an onFailure in the BayeuxClient.

This is because the long polling timeout in the LMS is 60 seconds and the default
timeout in LongPollingTransport is 10 seconds. There seem to be some code in the
server and client, which attempts to negotiate the timeout of the long polling
request (the connect message), but it doesn't work.

The timeout further caused a new connect request to be issued by the bayeux client,
leading to 2 outstanding connect messages (from the perspective of LMS), which is
not legal according to the Bayeux Protocol Specification.

The issue is fixed by setting the MAX_NETWORK_DELAY_OPTION to twice (120 sec.) the LMS
LONG_POLLING_TIMEOUT of 60 seconds.

Also abandon WebSocketTransport as LMS does not implement it.