Implement option to select Download storage
SDcard or internal

For devices and API levels which permits it.

For API level 21 (i.e. Android 5.0) it is (finally) possible to determine
whether a device has removable storage, which may be used for media, while
the standard public media storage is internal.

If this is the case the removeable storage can be selected in the preferences.
This storage differs from the external storage public directory, in that these
files are removed when the application is uninstalled, so the user is warned
about that.

This implements the reminder of issue #132