Handle type-unsafe responses from LMS
LMS will sometimes return integers as 2 and sometimes as "2". This is
pretty much unpredictable e.g.
$outbuf .= sprintf("%X", length($$body)) . $CRLF . $$body . $CRLF;
will send scalars as JSON strings, while
$outbuf .= $$body;
sends them as number values.
See <http://search.cpan.org/~mlehmann/JSON-XS-3.02/XS.pm#PERL_-%3E_JSON>
for details.

As we are not using customized parsers for our model objects, we simply
parse incoming records to map. This parsing guess the type of the elements,
so 2 becomes a Long, while "2" becomes a String.

To handle this we check the type of the element in the map befoire using it.

This involves changes all the model constructors to take a map of
<String, Object> instead of <String, String>.