Fix a crash if server names are 128-255 characters long.
The previous code converted a byte to an int directly. Bytes are signed,
so a length of 128 or above would be converted to a negative int, crashing
when the length was used later to extract a substring.

Now "& 0xFF" the value which avoids this.

Update the tests to include a couple of cases that catch this.

While I'm here, fix a couple of lint issues.

Fixes #226.