Allow the user to add/remove the current song from favorites.
Add a FavoriteButton widget. This displays the current state of the item
using two different vector drawables (using the support library). The widget
calls back to the hosting activity / fragment.

Host the button in layouts for NowPlayingFragment. On receipt of callback
messages call new SqueezeService methods to change the state on the server.

Change CliClient to handle the response and post a new FavoritesExists event
when the information is received.

Use a vector drawable for the button icon shown in the UI.