Implement a timeout for the server handshake
After connection Squeezer asks for server capabilities, and when done
notifies activities that handshake is complete. An indeterminate progressbar
is displayed while waiting for the handshake. If LMS does not reply to any of
these requests the progressbar stays on, and it is not possible to interact
with Squeezer.

With this commit if the handshake does not complete within the maximum time
(2 seconds), the connection is considered failed.

This does not fix the root cause of why such replies would be missing.