CometClient support command queries
i.e. commands with a "?" parameter.

Introduce a map of listeners to be notified when responses for the key in
the map entry comes in.

As the comet protocol do not echo the query also store the request along with
the response channel. To avoid state in the listeners the request parameters
are stored in the list of pending requests and passed to the listener, when
the response comes in, along with the response message. That way we can avoid
passing new instances of listeners to the bayeux library each time we publish
a message.

Initially support playerpref (for alarm preferences) and volume queries.