Handle interval in the handshake response
If the server sends an interval with the handshake response, set it to zero
so we can send the connect message immediately.
But store the interval so we can use it for subsequent connect messages, if
the server does not supply an interval in the connect response.

If the server doesn't send an interval with the connect response, but
we have one from the handshake response, then use that to avoid sending
connect messages continuously.