HomeActivity is now HomeMenuActivity and fetches
the menu for the active player from the server.

This provides access to all features from the server and new
features will be available in Squeezer automatically.

On the other hand this also looses some of the Specific Squeezer
features such as songs of album view and grid views.

Alarms and search functionality are still implemented as Squeezer
specific features.

This change also necessitate control of the active player and whether we
have an active player (any players connected to the server). This is
implemented by a new event ActivePlayerChanged which activities
subscribe to, sometimes in stead of the PlayersChanged event.