Fix idle disconnect on SN
The comet implementation on mysqueezebox.com is apparently different
than a local slimserver.

We use the cometd libraries which uses the connect message to keep the
connection alive. A connect is send, and resend using the interval in
the received advice. This is all handled by the base classes of the
cometd library, and works fine on local slimservers. On SN however, only
the first connect message is replied.

To fix this we do 2 things:
  1. Only send one connect message, to keep subsequent messages from
     expiring, as all messages expect a response.
  2. Set subscribe interval for the serverstatus, to keep the connection
     from disconnecting due to idle timeout, when there is no active
     player (in which case we would have a subscribe interval for the
     playerstatus).