Fix save current playlist
which had regressed because of the change to perform "do" action without
starting a new activity.

Since the input is collected by the newly started acitivity before the
action is performed, an activity is now started, if the do action
requires input.

Also add some javadoc to assist in preventing further regression.