Fix android:gravity=center bug on pre API 22
The gravity property does not work for items within layer-list on
certain API levels.

See for example here:
https://stackoverflow.com/questions/42885410/vector-drawable-in-layer-list-on-older-android-versions/43601921
or here:
https://stackoverflow.com/questions/34831142/how-to-center-vector-drawable-in-layer-list-without-scaling

These example shows that the issue can be worked around using a verbose
setup in each drawable. Since all embedded slim icons share the same background
and inset of the foreground icon, it is instead implemented in code.